// src/lib/hafas-types.ts
export interface components {
	schemas: {
// ...
		JourneyDetail: {
			TechnicalMessages?: components["schemas"]["TechnicalMessages"];
			Warnings?: components["schemas"]["Warnings"];
			serverVersion?: string;
			dialectVersion?: string;
			version?: string;
			/** Format: date-time */
			planRtTs?: string;
			errorCode?: string;
			errorText?: string;
			internalErrorCode?: string;
			internalErrorText?: string;
			internalErrorTextOut?: string;
			requestId?: string;
			Stops: components["schemas"]["Stops"];
			Names?: components["schemas"]["Names"];
			Product?: components["schemas"]["ProductType"][];
			Directions?: components["schemas"]["Directions"];
			Notes?: components["schemas"]["Notes"];
			Messages?: components["schemas"]["Messages"];
			/** @enum {string} */
			JourneyStatus?: "P" | "R" | "A" | "S";
			Polyline?: components["schemas"]["Polyline"];
			PolylineGroup?: components["schemas"]["PolylineGroup"];
			ServiceDays?: components["schemas"]["ServiceDays"][];
			referencedJourney?: components["schemas"]["ReferencedJourneyType"][];
			lastPos?: components["schemas"]["Coordinate"];
			/** Format: date-time */
			lastPosReported?: string;
			/** Format: int32 */
			lastPassRouteIdx?: number;
			/** Format: int32 */
			lastPassStopRef?: number;
			/** Format: int32 */
			rtLastPassRouteIdx?: number;
			/** Format: int32 */
			rtLastPassStopRef?: number;
			ParallelJourneyRef?: components["schemas"]["ParallelJourneyRefType"][];
			cancelled?: boolean;
			partCancelled?: boolean;
			reachable?: boolean;
			redirected?: boolean;
			uncertainDelay?: boolean;
			dayOfOperation: string;
			ref: string;
		};
// ...
	};
	responses: never;
	parameters: never;
	requestBodies: never;
	headers: never;
	pathItems: never;
}